+++ CHANGELOG.md
@@ -7,15 +7,32 @@ To learn more, we recommend reading [this document](README.md).
 ## History
 
 - [Unreleased](#unreleased)
+- [v2.0.0](#v200)
 - [v1.0.0](#v100)
 
 ## Unreleased / under development
 
 ### New
--
+- Refactoring of resources installations
+- Path-aware definitions evaluations
 
 ### Other
 -
+
+## v2.0.0
+
+### New
+- General codebase refactoring
+- Refactoring of functions and install mechanisms
+- General bug fixing
+- Support for Docker Engine
+- New interactive interface
+- Added support for kubernetes cluster reapplication
+- Added support for docker/podman on-demand image building
+
+### Other
+- Implementation of Strimzi
+- Implementation of Mayfly
 
 ## v1.0.0